requirements.txt
websockets==12.0
requests==2.31.0
customtkinter==5.2.2
pystray==0.19.5
Pillow==10.2.0
asyncio
